Rewrite NextPrezlyApi to build on top of `ContentDelivery` client
